Move config related functions into the config class
